fix: move default_url_options to controller extension (#2780)
