+++ app/components/avo/sidebar_component.html.erb
@@ -25,7 +25,7 @@
 
               <div class="w-full space-y-1">
                 <% dashboards.sort_by { |r| r.navigation_label }.each do |dashboard| %>
-                  <%= render Avo::Sidebar::LinkComponent.new label: dashboard.navigation_label, path: Avo::Dashboards::Engine.routes.url_helpers.dashboard_path(dashboard) %>
+                  <%= render Avo::Sidebar::LinkComponent.new label: dashboard.navigation_label, path: helpers.avo_dashboards.dashboard_path(dashboard) %>
                 <% end %>
               </div>
             </div>
