+++ app/controllers/avo/application_controller.rb
@@ -7,6 +7,7 @@ module Avo
     end
 
     include Avo::InitializesAvo
+    include Avo::CommonController
     include Avo::ApplicationHelper
     include Avo::UrlHelpers
     include Avo::Concerns::Breadcrumbs
@@ -297,24 +298,6 @@ module Avo
       I18n.with_locale(locale, &action)
     end
 
-    def default_url_options
-      result = super.dup
-
-      if params[:force_locale].present?
-        result[:force_locale] = params[:force_locale]
-      end
-
-      extra_options = get_extra_default_url_options
-
-      if extra_options.present?
-        extra_options.each do |param_name|
-          result[param_name] = params[param_name]
-        end
-      end
-
-      result
-    end
-
     def set_sidebar_open
       value = cookies["#{Avo::COOKIES_KEY}.sidebar.open"]
       @sidebar_open = value.blank? || value == "1"
@@ -345,16 +328,6 @@ module Avo
     end
 
     private
-
-    def get_extra_default_url_options
-      block_or_array = Avo.configuration.default_url_options
-
-      if block_or_array.respond_to?(:call)
-        instance_eval(&block_or_array)
-      else
-        block_or_array
-      end
-    end
 
     def choose_layout
       if turbo_frame_request?
